Move duplicate fixture into dedicated file
